requirements.txt
# django packages
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0

# deduce packages
pyspark==3.5.5
deduce==3.0.3
pandas==2.2.3
pyarrow==19.0.1

# needed for 'distutils'
setuptools==77.0.3
